Extracted asset path logic from core.py into a new asset path module.
